Added Service functionalities

+++ src/main/java/se/ifmo/ru/soa_2023_lab2_service1/storage/repository/api/TicketRepository.java
@@ -17,7 +17,7 @@ public interface TicketRepository {
 
     TicketEntity getMinimumType();
 
-    long countTicketByPrice(int price);
+    long countTicketByPrice(float price);
 
     List<TicketEntity> getTicketGreaterType(String ticketType);
